Support handling of CHECK constraints correctly
Before this patch, when the target migration included a CHECK constraint
and the migration was applied multiple times, it failed due to a duplicated
constraint. This happened because the migration always executed the
`ALTER TABLE ... ADD CONSTRAINT ...` operation in the unidirection each time.

This patch fixes the issue by ensuring that add_check_constraint is dumped
correctly, preventing such duplication errors.

--

FYI:

However, Ridgepole still always attempts to execute ALTER TABLE for each
CHECK constraint—dropping and then re-adding it.

e.g.,

```
\# ALTER TABLE `mytable` DROP CHECK value_range
\# ALTER TABLE `mytable` ADD CONSTRAINT value_range CHECK (value > 0 AND value <= 1)
```

This behavior is consistent with the original schema dumper, so it should not be
considered a bug in the dumper itself.

Signed-off-by: moznion <moznion@mail.moznion.net>